Split up large chunks of untested code into very rough submodules. No unit tests. What can go wrong?
