+++ ti/datasources/json.py
@@ -1,0 +1,22 @@
+import json
+from os import path
+
+class JsonStore(object):
+
+    def __init__(self, filename):
+        self.filename = filename
+
+    def load(self):
+
+        if path.exists(self.filename):
+            with open(self.filename) as f:
+                data = json.load(f)
+
+        else:
+            data = {'work': [], 'interrupt_stack': []}
+
+        return data
+
+    def dump(self, data):
+        with open(self.filename, 'w') as f:
+            json.dump(data, f, separators=(',', ': '), indent=2)
